+++ Example/SwiftDemo/SwiftDemo/AppDelegate.swift
@@ -97,11 +97,13 @@ class AppDelegate: UIResponder, UIApplicationDelegate
         }
         else
         {
-            let idToken: String = user.authentication.accessToken
+            let googleToken: String = user.authentication.accessToken
+            let refreshToken: String = user.authentication.refreshToken
+            let clientID: String = user.authentication.clientID
             if let navVC = self.window?.rootViewController as? UINavigationController,
             let currentVC = navVC.topViewController
             {
-                LoginRadiusSocialLoginManager.sharedInstance().convertGoogleToken(toLRToken: idToken, in:currentVC, completionHandler: {( data ,  error) -> Void in
+                LoginRadiusSocialLoginManager.sharedInstance().convertGoogleToken(toLRToken: googleToken,google_refresh_token:refreshToken, google_client_id:clientID, in:currentVC, completionHandler: {( data ,  error) -> Void in
                 NotificationCenter.default.post(name: Notification.Name("userAuthenticatedFromNativeGoogle"), object: nil, userInfo: ["data":data as Any,"error":error as Any])
 
                 })